Added lines 15-18 to the clean function
This code is meant to remove the vertical bar, because this character ruins our data. Whenever one of the lines from the original document split on a word, EEBO marked this not as a hyphen but as a vertical bar. This happened a lot, in all sorts of random places. So we have a lot of nonsensical words in our hashtables.